Invalidate sessions when password is changed
Store a session key in the session that we regenerate every time
the password changes. If they don't match then the session is
invalid.

If the user changes their own password, we'll update the current
session with the new key but any others will invalidate.

If an admin changes a password via /wiki/users, all sessions for
that user will be invalidated.

Closes #75