Make recent changes way more performant/useful
Instead of trying to use the stats for the changes (which
read and diff every file involved), just bite the bullet and
manually compare the trees ourselves.

This speeds the process up from taking ~a second on my desktop
to being unnoticably fast. It also now correctly shows diffs for
binary files (images/site config/etc).